% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anonCols}
\alias{anonCols}
\title{anonCols
Anonymise the values in one more columns of a data frame
function from: \href{https://stackoverflow.com/a/10455661}{Stack Overflow}}
\usage{
anonCols(df, colIDs)
}
\arguments{
\item{df}{The data frame name}

\item{colIDs}{The location of the columns with values to anonymise. Use either column names or integer indices as shownin the two examples.}
}
\description{
anonCols
Anonymise the values in one more columns of a data frame
function from: \href{https://stackoverflow.com/a/10455661}{Stack Overflow}
}
\examples{
anonCols(datasets::sleep, c("group","ID"))
anonCols(datasets::sleep, c(2,3))
}
\keyword{anonymise}
